requirements.txt
torch
torchvision
torchaudio
pandas
openpyxl
scikit-learn
pyarrow
pyyaml
scipy
numpy
